changed types the report works with

+++ src/django/sqls/0150_revert_add_percent_to_number.sql
@@ -1,1 +1,1 @@
-DROP FUNCTION IF EXISTS add_percent_to_number(int8, int8);
+DROP FUNCTION IF EXISTS add_percent_to_number(numeric, numeric);
